+++ .idea/misc.xml
@@ -3,6 +3,7 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="../../../../../layout/compose-model-1627897379997.xml" value="0.19087837837837837" />
         <entry key="app/src/main/res/layout/activity_second.xml" value="0.11322463768115942" />
       </map>
     </option>
