+++ app/src/main/java/dpm/android/with/kotlin/lifecycle/compose/utils/Color.kt
@@ -1,0 +1,9 @@
+
+package dpm.android.with.kotlin.lifecycle.compose.utils
+
+import androidx.compose.ui.graphics.Color
+
+val purple200 = Color(0xFFBB86FC)
+val purple500 = Color(0xFF6200EE)
+val purple700 = Color(0xFF3700B3)
+val teal200 = Color(0xFF03DAC5)
