+++ app/src/main/java/dpm/android/with/kotlin/lifecycle/compose/utils/Shape.kt
@@ -1,0 +1,11 @@
+package dpm.android.with.kotlin.lifecycle.compose.utils
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.Shapes
+import androidx.compose.ui.unit.dp
+
+val shapes = Shapes(
+    small = RoundedCornerShape(4.dp),
+    medium = RoundedCornerShape(4.dp),
+    large = RoundedCornerShape(0.dp)
+)
