+++ app/src/main/java/dpm/android/with/kotlin/lifecycle/compose/utils/Type.kt
@@ -1,20 +1,4 @@
-/*
- * Copyright 2020 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package dpm.android.with.kotlin.lifecycle.compose
+package dpm.android.with.kotlin.lifecycle.compose.utils
 
 import androidx.compose.material.Typography
 import androidx.compose.ui.text.TextStyle
