+++ app/src/main/java/dpm/android/with/kotlin/lifecycle/old/MainActivity.kt
@@ -1,4 +1,4 @@
-package dpm.android.with.kotlin.lifecycle
+package dpm.android.with.kotlin.lifecycle.old
 
 import android.content.Intent
 import android.os.Bundle
