+++ app/src/main/java/dpm/android/with/kotlin/lifecycle/old/SecondActivity.kt
@@ -1,4 +1,4 @@
-package dpm.android.with.kotlin.lifecycle
+package dpm.android.with.kotlin.lifecycle.old
 
 import android.os.Bundle
 import android.util.Log
