+++ app/src/main/res/layout/activity_second.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".lifecycle.SecondActivity">
+    tools:context=".lifecycle.old.SecondActivity">
 
     <TextView
         android:id="@+id/textView2"
